Move SingleArg to the args module
