+++ metriken/derive/src/args.rs
@@ -2,6 +2,7 @@
 // Licensed under the Apache License, Version 2.0
 // http://www.apache.org/licenses/LICENSE-2.0
 
+use proc_macro2::TokenStream;
 use quote::ToTokens;
 use std::fmt::{Display, Formatter, Result};
 use syn::parse::{Parse, ParseStream};
@@ -51,5 +52,35 @@ impl Display for ArgName {
             Ident(ident) => ident.fmt(f),
             Crate(_) => f.write_str("crate"),
         }
+    }
+}
+
+/// A single argument to an attribute macro.
+///
+/// ```text
+/// #[macro(name = value, a = "string")]
+///         ^^^^^^^^^^^^  ^^^^^^^^^^^^
+/// ```
+pub(crate) struct SingleArg<T> {
+    pub ident: ArgName,
+    pub eq: Token![=],
+    pub value: T,
+}
+
+impl<T: Parse> Parse for SingleArg<T> {
+    fn parse(input: ParseStream) -> syn::Result<Self> {
+        Ok(Self {
+            ident: input.parse()?,
+            eq: input.parse()?,
+            value: input.parse()?,
+        })
+    }
+}
+
+impl<T: ToTokens> ToTokens for SingleArg<T> {
+    fn to_tokens(&self, tokens: &mut TokenStream) {
+        self.ident.to_tokens(tokens);
+        self.eq.to_tokens(tokens);
+        self.value.to_tokens(tokens);
     }
 }
